Pong pt 1
Basic if broken functionality. Score increments but is not global (that is each goal doesn't keep track of the other goal's score) and refuses to reset to 0 - 0 properly at end of game. Because of Ball Prefabs shooting off through the wall after two iterations for some reason, elected to make a "permanent" ball that ricochets off the back wall as a temporary fix. Hard coded starting values so that ball will ricochet properly off the paddle. Paddle ricochet is non functional (i.e. ball can't be controlled only reflected).